MagLDA-PyTorch/requirements.txt
matplotlib
numpy
pandas
pillow
scipy
seaborn
setuptools
sklearn
visdom
pytorch
